requirements.txt
plyer==2.1.0
rapidocr-onnxruntime
opencv-python==4.12.0.88
pillow==12.0.0
pyscreeze==1.0.1
pyautogui==0.9.54
pynput==1.8.1
psutil==7.1.3
rich==14.2.0
loguru==0.7.3
cryptography>=42.0.0
nuitka
pyl10nc

# Windows 专用
pygetwindow==0.0.9; sys_platform == "win32"
pywin32; sys_platform == "win32"

# Linux 专用
ewmh>=0.1.6; sys_platform == "linux"
